Implement flow for string module
This was mostly straightforward, however, it required a bit of refactoring for the `contains` function. There is a call to `.map` on whatever the `patterns` variable gets set to, and though it's easy to see that it will always get casted to an array, `flow` apparently can't detect this, so it'd throw a type checking error saying that ".map is not a String method".

Breaking it up with the `do_contains` function did the trick, since at that point `patterns` will definitely be a variable. I made it a function outside of the exported module as to keep it a private function, but if we want to go about that a different way, I'm game.