Fix and simplify `contains` code
Apparently my previous assumption wasn't correct. Via this [this GitHub issue](https://github.com/facebook/flow/issues/3948) I determined that `flow` seems to be able to trust certain control flows, like conditionally switching on a call to `Array.isArray` (but doesn't trust `lodash`). This makes it so there is no need for the private function, and we can leave the function mostly alone.